Time: 3 ms (18.2%), Space: 8.7 MB (14.89%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.cpp
@@ -7,8 +7,8 @@ public:
         int r=x;
         int ans = -1;
         while(l<=r){
-            long long mid=l+(r-l)/2;
-            long long mul=mid*mid;
+            int mid=l+(r-l)/2;
+            long long mul=(long long)mid*mid;
             if(mul<=x){
                 ans = mid;
                 l = mid+1;
